+++ meta/alias.ts
@@ -1,18 +1,21 @@
 import { dirname, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
+import type { Alias } from 'vite';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
-export const alias = {
-  'mixte': resolve(__dirname, '../packages/mixte/index'),
-  '@mixte/use/register': resolve(__dirname, '../packages/use/src/register'),
-  '@mixte/use': resolve(__dirname, '../packages/use/index'),
-  '@mixte/components': resolve(__dirname, '../packages/components/index'),
-};
+export const alias: Alias[] = [
+  { find: 'mixte', replacement: resolve(__dirname, '../packages/mixte/index') },
+  { find: '@mixte/use/register', replacement: resolve(__dirname, '../packages/use/src/register') },
+  { find: '@mixte/use', replacement: resolve(__dirname, '../packages/use/index') },
+  { find: '@mixte/components', replacement: resolve(__dirname, '../packages/components/index') },
+  { find: /^@mixte\/snippets\/(.*)$/, replacement: resolve(__dirname, '../packages/snippets/src/$1/index') },
+];
 
-export const testAlias: Record<keyof typeof alias, string> = {
-  'mixte': resolve(__dirname, '../packages/mixte/dist/index'),
-  '@mixte/use/register': resolve(__dirname, '../packages/use/dist/register'),
-  '@mixte/use': resolve(__dirname, '../packages/use/dist/index'),
-  '@mixte/components': resolve(__dirname, '../packages/components/dist/index'),
-};
+export const testAlias: Alias[] = [
+  { find: 'mixte', replacement: resolve(__dirname, '../packages/mixte/dist/index') },
+  { find: '@mixte/use/register', replacement: resolve(__dirname, '../packages/use/dist/register') },
+  { find: '@mixte/use', replacement: resolve(__dirname, '../packages/use/dist/index') },
+  { find: '@mixte/components', replacement: resolve(__dirname, '../packages/components/dist/index') },
+  { find: /^@mixte\/snippets\/(.*)$/, replacement: resolve(__dirname, '../packages/snippets/dist/$1') },
+];